+++ jsonrpc-core/src/main/java/com/limehee/jsonrpc/core/JsonRpcInterceptorExecutionException.java
@@ -1,0 +1,8 @@
+package com.limehee.jsonrpc.core;
+
+public final class JsonRpcInterceptorExecutionException extends RuntimeException {
+
+    public JsonRpcInterceptorExecutionException(Throwable cause) {
+        super(cause);
+    }
+}